rest api
#jersey is added
#swagger is added
#validation is added
#payment services are added

+++ src/main/java/com/freimanvs/company/servlets/EmployeeServlet.java
@@ -33,30 +33,49 @@ public class EmployeeServlet extends HttpServlet {
 
         String id = req.getParameter("id");
 
+        String response;
+
         //get all
         if (id == null) {
 
             List<Employee> employees = employeeService.getList();
-            String jsonString = JSONB.toJson(employees);
+            response = JSONB.toJson(employees);
 
+            resp.setContentLength(response.length());
             resp.setContentType("application/json");
+            resp.setCharacterEncoding("utf8");
+            resp.setStatus(200);
             try (PrintWriter pw = resp.getWriter()) {
-                pw.print(jsonString);
+                pw.print(response);
                 pw.flush();
             }
         }
 
         //get one by id
         else {
+            Employee employee = employeeService.getById(Long.parseLong(id));
 
+            if (employee == null) {
+                notFound(resp);
+            } else {
+                response = JSONB.toJson(employee);
+                resp.setStatus(200);
+                resp.setContentType("application/json");
+                resp.setCharacterEncoding("utf8");
+                resp.setContentLength(response.length());
+                try (PrintWriter pw = resp.getWriter()) {
+                    pw.print(response);
+                    pw.flush();
+                }
+            }
         }
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String contentType = req.getContentType();
-        if (!contentType.equals("application/json")) {
-            resp.sendError(400, "application/json is required");
+        if (!contentType.contains("application/json")) {
+            badRequest(resp);
             return;
         }
 
@@ -70,8 +89,10 @@ public class EmployeeServlet extends HttpServlet {
 
             long id = employeeService.add(emp);
 
-//            Employee empFromDB = employeeService.getById(id);
-//            System.out.println(JSONB.toJson(empFromDB));
+            String absolute_path = req.getProtocol().split("[/]")[0].toLowerCase() + "://" + req.getLocalAddr() +
+                    ":" + req.getLocalPort() + req.getRequestURI() + "?id=" + id;
+            resp.setHeader("location", absolute_path);
+            resp.setStatus(201);
         }
     }
 
@@ -80,20 +101,32 @@ public class EmployeeServlet extends HttpServlet {
 
         String contentType = req.getContentType();
         if (!contentType.startsWith("application/json")) {
-            resp.sendError(400, "application/json is required");
+            badRequest(resp);
             return;
         }
 
         long id_param = Long.valueOf(req.getParameter("id"));
-        try (InputStream is = req.getInputStream();
-             PrintWriter pw = resp.getWriter()) {
-            StringBuilder json = new StringBuilder();
-            while (is.available() != 0) {
-                json.append((char)is.read());
+
+        Employee employee = employeeService.getById(id_param);
+        if (employee == null) {
+            notFound(resp);
+        } else {
+
+            try (InputStream is = req.getInputStream();
+                 PrintWriter pw = resp.getWriter()) {
+                StringBuilder json = new StringBuilder();
+                while (is.available() != 0) {
+                    json.append((char) is.read());
+                }
+                Employee emp = JSONB.fromJson(json.toString(), Employee.class);
+
+                employeeService.updateById(id_param, emp);
+
+                String absolute_path = req.getProtocol().split("[/]")[0].toLowerCase() + "://" +
+                        req.getLocalAddr() + ":" + req.getLocalPort() + req.getRequestURI() + "?id=" + id_param;
+                resp.setHeader("location", absolute_path);
+                resp.setStatus(201);
             }
-            Employee emp = JSONB.fromJson(json.toString(), Employee.class);
-            employeeService.updateById(id_param, emp);
-//            System.out.println(JSONB.toJson(employeeService.getById(id_param)));
         }
 
     }
@@ -101,7 +134,38 @@ public class EmployeeServlet extends HttpServlet {
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         long id_param = Long.valueOf(req.getParameter("id"));
+
+        Employee employee = employeeService.getById(id_param);
+
+        if (employee == null) {
+            notFound(resp);
+            return;
+        }
+
         employeeService.deleteById(id_param);
-//        resp.sendRedirect("/login");
+        resp.setStatus(204);
+    }
+
+    private static void notFound(HttpServletResponse resp) throws IOException {
+        String response = "HTTP 404 Not Found";
+        resp.sendError(404);
+        resp.setContentType("text/plain");
+        resp.setCharacterEncoding("utf8");
+        resp.setContentLength(response.length());
+        resp.setHeader("connection", "close");
+
+        try (PrintWriter pw = resp.getWriter()) {
+            pw.print(response);
+            pw.flush();
+        }
+    }
+
+    private static void badRequest(HttpServletResponse resp) throws IOException {
+        String response = "application/json is required";
+        resp.sendError(400, response);
+        resp.setContentType("text/plain");
+        resp.setCharacterEncoding("utf8");
+        resp.setContentLength(response.length());
+        resp.setHeader("connection", "close");
     }
 }
